Reorganize custom event listeners
