+++ source/features/table-input.tsx
@@ -7,8 +7,8 @@ import * as pageDetect from 'github-url-detection';
 import * as textFieldEdit from 'text-field-edit';
 
 import features from '.';
-import onCommentEdit from '../github-events/on-comment-edit';
 import smartBlockWrap from '../helpers/smart-block-wrap';
+import {onCommentEdit} from '../github-events/on-fragment-load';
 
 function addTable({delegateTarget: square}: delegate.Event<MouseEvent, HTMLButtonElement>): void {
 	const field = square.form!.querySelector('textarea')!;
